Ruma Clan Lore Rework v2
Changes Kashira to Seonjang

Changes Isha to Sasu

Changes Kyodai to Gun-in

The Ruma clan is A band of foreign outcast Kazengunites. The Ruma Clan were outcasts from the Xinyi Dynasty, believed to be associated with the rebels at the time. The clan departed lest they risked being executed for such suspicions, or worse. It is no organized group of soldiers, but rather a loose collection of experienced fighters.

Swords and tattoos were changed as well.