Headshot in Chat initial push
This will be a PR to allow headshots to be down in chat. its an option right below the headshot button and will show the headshot in 100x100.

Based on Helmsdeep/dreamkeep's idea, I decided to add this in as an option. it can be turned on or off before you join the game.