Chosen Redefined & Vale Edits
A few renames slipped through, that, well, really shouldn't have.
This corrects that.

Additionally, it modifies the trait the Priest/Bishop gets to reference the Ten as a total. As opposed to just Astrata. They retain the Astratan robes, of course, as that's one aspect of their image. But they're not just Astrata's chosen by way of fluff now. God, I hate the Undivided change.