Clergy Titles
Simply returns titles from RW.
As opposed to 'Father/Mother' for the Priest, we've adopted Prelate as a universal title for Bishop. It should be fine? I guess?

Otherwise, Templars and Acolytes are titled 'Brother/Sister', as they were in RW. This title is already used in game, but it's flavourful to prepend it to character names.