Translations | De-Miracle | Death's Door
Cleans up the Xylix file. Returns the translations.

Removes Cantor's guaranteed mockery, given you got miracles anyhow and it could double a Xylixian's mockery slots. You've given up your life as a proper bard. Suffer for it.

Returns Death's Door to Necrans. I've never seen the crash, or supposed issues to come with it. It was an excellent tool. The one thing I HAVE seen come of it, a long while ago, had been the monochrome filter refusing to drop when required, but that's a simple fix if it persists even now. If it causes crashing, it's probably mob client garbage(or lack thereof) and stupid easy to fix. But, again, I've not seen this, and I've not had such reported to me. So, fingers crossed this is no longer an issue. I'll do a double take if it is and eat a shoe. (I'll probably have to eat that shoe, but this should be fine.)