Hallowed Ground vs Gilded Foes
Applies a tremendous debuff to folks in ritual armour when trying to enter hallowed ground. That being, as of now, just the cathedral. This doesn't touch anyone else. Only those trying to enter while wearing gilded/vicious/darksteel equipment by way of the rites, which provides a trait upon successfully working now to the target.

The trait itself marks them as an overt heretic, for the purposes of code stuff. It is not player facing. All it does is provide the debuff when walking on hallowed ground, which is their visible warning by that point.

Additionally, it removes an improper marking from the druid's outer grove areas. It's not intended to be hallowed, which also provides a moodbuff to Undivided patron Tennites as an aside.