So much Gardening
- fixed a lot of mapping errors from the last Mapping Changes
- Worked extensively on the southern forests and some of the northern woods, too.

Mostly: Ruins; Woodlands, grass tiles and enhancement of nature as I always found the mapgen by itself to be a bit bleak at making unique areas.

In the south, most of the buildings near the wardens path were moved towards the bog entrance; giving it a bit of a dilapidated shanty town vibe with spooky scary skeletons.

I worked to make the sorroundings more boggish, as if the landscape doesn't totally adhere to sections.

The wardens camp was enhanced, whilst the timber yard was moved to encourage a good logging area that wasn't right next to a leyline.

I also worked on changing the traveltile locations for the bandits; giving them an area in the mountains and one in the swamp so both locations were more fitting toward longterm pursuits. These will probably change according to whatever is requested as they might be a bit too far for the current pop.