Wretch Adjustments
Drops Wretch from 8 slots, to 3. We don't have the pop or want as of now to have so many slots for such powerful classes.
Removes their universal +1 to stats across the board. This was a holdover from when they got a buff for not playing WW/Vampire elsewhere. We don't need this. They're plenty strong as is.