Bed Breaker & Good Lover Messages
I hate you people. I could not begin to describe just how much I despise this community for asking me to return this.
That said, it's for parity with OldRW.
What do you think this does?

If it's not clear, it adds Bed Breaker to WWs, Brigands, the Dungeoneer, Giant virtue and Enlarge spell.
Also includes messages for Good Lover, to indicate someone has it without triumph generation.